+++ src/test/java/co/fineants/api/domain/member/service/SignupVerificationServiceTest.java
@@ -1,5 +1,8 @@
 package co.fineants.api.domain.member.service;
 
+import static org.assertj.core.api.Assertions.*;
+
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentMatchers;
@@ -8,6 +11,7 @@ import org.mockito.internal.verification.Times;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import co.fineants.AbstractContainerBaseTest;
+import co.fineants.api.global.errors.exception.business.VerifyCodeInvalidInputException;
 import co.fineants.api.infra.mail.EmailService;
 import jakarta.mail.internet.MimeMessage;
 
@@ -18,6 +22,9 @@ class SignupVerificationServiceTest extends AbstractContainerBaseTest {
 
 	@Autowired
 	private EmailService mockEmailService;
+
+	@Autowired
+	private VerifyCodeRepository verifyCodeRedisRepository;
 
 	@DisplayName("이메일로 회원가입 검증 코드 전송 테스트")
 	@Test
@@ -30,6 +37,33 @@ class SignupVerificationServiceTest extends AbstractContainerBaseTest {
 		BDDMockito.then(mockEmailService)
 			.should(new Times(1))
 			.sendEmail(ArgumentMatchers.any(MimeMessage.class));
+	}
+
+	@DisplayName("사용자는 검증코드를 제출하여 검증코드가 일치하는지 검사한다")
+	@Test
+	void checkVerifyCode() {
+		// given
+		String email = "dragonbead95@naver.com";
+		String code = "123456";
+		verifyCodeRedisRepository.save(email, code);
+		// when & then
+		Assertions.assertDoesNotThrow(() -> service.verifyCode(email, code));
+	}
+
+	@DisplayName("사용자는 매치되지 않은 검증 코드를 전달하며 검사를 요청했을때 예외가 발생한다")
+	@Test
+	void checkVerifyCode_whenNotMatchVerifyCode_thenThrowException() {
+		// given
+		String email = "dragonbead95@naver.com";
+		String code = "234567";
+
+		// when
+		Throwable throwable = catchThrowable(() -> service.verifyCode(email, code));
+
+		// then
+		assertThat(throwable)
+			.isInstanceOf(VerifyCodeInvalidInputException.class)
+			.hasMessage("234567");
 	}
 
 }
