+++ src/test/java/co/fineants/config/TestConfig.java
@@ -17,7 +17,6 @@ import co.fineants.api.domain.fcm.service.FirebaseMessagingService;
 import co.fineants.api.domain.kis.client.KisClient;
 import co.fineants.api.domain.kis.repository.FileHolidayRepository;
 import co.fineants.api.domain.kis.service.KisService;
-import co.fineants.api.domain.member.service.VerifyCodeRepository;
 import co.fineants.api.domain.portfolio.service.PortfolioCacheService;
 import co.fineants.api.global.common.delay.DelayManager;
 import co.fineants.api.global.common.time.LocalDateTimeService;
@@ -25,9 +24,6 @@ import co.fineants.api.infra.mail.EmailService;
 
 @TestConfiguration
 public class TestConfig {
-	@MockBean
-	private VerifyCodeRepository verifyCodeRedisRepository;
-
 	@MockBean
 	private EmailService emailService;
